COMPLETE=========================================================== BUILD 2018.0000004 -changed Flame and Stab hitbox to use ForwardRectHitBox -fixed ability levelup issue -fixed Inventory.scroll items and discard -TONS of Refactoring bits -add multipliers next to stats that are affected on ability select screen --make file that inputs all the Ability detail text -complete refactor of Anim class -fix drop animation on hold key
WORKING ON==========================================================
-fixing compiler warnings

LOW PRIORITY=======================================================
-drop spellbook when max attack bar is exceed

BUGS============================================================
-corrupted player stats in highscorestate
-shooting the left edge breaks hitbox placement
-hitbox doesnt always show on screen
-screen wrapping breaks occasionally
-items in shop inventory default to boxes after sell

+++ core/src/com/quadx/dungeons/Damage.java
@@ -47,7 +47,7 @@ public class Damage {
         c = (float) m.getPower();
         d = (float) m.getLevel();
 
-        damage = (int) (Math.round(((c*d)+(b*8))/(a)));
+        damage =(Math.round(((c*d)+(b*8))/(a)));
         if (damage < 0) //checks for negative damage
             damage = defaultDamage;
         if (rn.nextFloat() < .1) {
